Initial wrapping of collapseTracklets.
The difficulty with Tracklets is that Tracklets.indices is a stl vector, so we
can see it from python but we can't edit it. We could make it opaque to
pybind11, as with TrackletSet, but that requires defining a new class which will
have to replace all uses of std::vector<unsigned int>. In this case I think the
easiest option is to make the Tracklet constructor accept a list of indicies,
then replace the indices member with a method that returns the python set. This
discourages anyone from attempting to edit the indices set.