Differentiate between det ID and index.
Much of the tracklet code assumes that a detection's ID parameter is also its
position in the detection vector. This is very fragile and makes the ID
parameter useless for storing the IDs generated by the DIA Source catalog.

To get around that, findTracklets() now uses the detection's index parameter,
which is initialized by default to the detection ID. If one uses IDs from an
external source, then the index parameter of all detections must be reset to its
position in the detection vector prior to calling findTracklets().

This is all very annoying, and a much better solution will be to make tracklets
contain detection IDs rather than indices. This will require substantial changes
to implement.